+++ packages/jest-util/src/buffered_console.js
@@ -15,6 +15,7 @@ import type {
   LogTimers,
 } from 'types/Console';
 
+import assert from 'assert';
 import {Console} from 'console';
 import {format} from 'util';
 import chalk from 'chalk';
@@ -63,8 +64,10 @@ export default class BufferedConsole extends Console {
   }
 
   assert(...args: Array<any>) {
-    if (args[0]) {
-      this._log('assert', format(...args.slice(1)));
+    try {
+      assert(...args);
+    } catch (error) {
+      this._log('assert', error.toString());
     }
   }